+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\IO\Streams\Stream\Write::sinkAttempts()`
+- `Innmind\IO\Sockets\Clients\Client::sinkAttempts()`
+
 ## 3.2.0 - 2025-05-09
 
 ### Added